pyproject.toml
[project]
name = "futureshots"
version = "0.1.0"
description = "A diary for smart and lazy street photogs"
requires-python = ">=3.10"

[tool.poetry]
name = "futureshots"
version = "0.1.0"
description = "A diary for smart and lazy street photogs"
authors = ["pyrocat"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.3"
djangorestframework = "^3.13.1"
knox = "0.1.14"
pillow = "^10.0.0"
exif = "^1.6.0"
drf-nested-routers = "^0.93.4"
django-mptt = "^0.15.0"
pydantic = "^2.4.2"
django-filter = "^23.4"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
